Enrich the error messages with the bound names that are available.
The user often do not know whether it's not the correct name, or whether it was not defined, etc. It's easier to get this information when debugging.
